git-svn-id: https://bsisvn.mayo.edu/main/projects/g/Genome_GPS/trunk@2949 8f329c14-d232-4f17-8684-7ee34322b8dc

+++ filter_variant_vqsr.sh
@@ -124,8 +124,11 @@ else
 				check=`[ -s $output/temp/$input_name.tranches.pdf ] && echo "1" || echo "0"`
 				if [ $check -eq 0 ]
 				then
-					rm `grep -l $output/plot/$input_name.plots.R *.log`
-					rm core.*
+					if [[  `find . -name '*.log'` ]]
+					then
+						rm `grep -l $output/plot/$input_name.plots.R *.log`
+						rm core.*
+					fi
 				fi
 				let count=count+1
 			done
@@ -215,8 +218,11 @@ else
 				check=`[ -s $output/temp/$input_name.tranches.pdf ] && echo "1" || echo "0"`
 				if [ $check -eq 0 ]
 				then
-					rm `grep -l $output/plot/$input_name.plots.R *.log`
-					rm core.*
+					if [[  `find . -name '*.log'` ]]
+					then
+						rm `grep -l $output/plot/$input_name.plots.R *.log`
+						rm core.*
+					fi
 				fi
 				let count=count+1
 			done
